chore(test): honest coverage scope + extract harness into its own file
(a) codecov.yml: ignore the genuinely-untestable-headlessly paths (binary
entry main.rs; the crossterm event loop tui/mod.rs which reads the real TTY)
so the reported % means "of code we CAN test headlessly, X% is tested" rather
than being dragged down by paths no test can reach. Not chasing 100% — the
remaining gap is the event loop (needs injectable events) + defensive/OS-
boundary code, where tests turn brittle or impossible.

(b) Split tui_renderer.rs (~1760 lines, half tests) into tui_renderer/mod.rs
(production) + tui_renderer/harness.rs (the 38 headless integration tests).
The harness is a `#[cfg(test)] mod harness;` submodule of tui_renderer, so it
keeps `use super::*` and full crate-internal access to the #[cfg(test)] seams
(floor_motion/floor_buf/inject_coffee/frame_buffer) and private consts — no
API widening, just separation of harness tests from production code. Unit
tests stay in `#[cfg(test)] mod tests` next to their code; cross-crate
contract tests stay in tests/.

561 tests, clippy -D warnings + fmt clean.
